+++ test/client.spec.ts
@@ -1,7 +1,6 @@
 import { readFileSync } from "fs";
 import { join } from "path";
-import { deepEqual, equal } from "assert";
-import { Readable } from "stream";
+import { deepEqual, equal, ok } from "assert";
 import Client from "../lib/client";
 import * as Types from "../lib/types";
 import { getStreamData } from "./helpers/stream";
@@ -11,7 +10,6 @@ const TEST_PORT = parseInt(process.env.TEST_PORT, 10);
 
 const client = new Client({
   channelAccessToken: "test_channel_access_token",
-  channelSecret: "test_channel_secret",
 });
 
 const getRecentReq = (): any =>
@@ -54,7 +52,7 @@ describe("client", () => {
     equal(req.method, "POST");
     equal(req.body.replyToken, "test_reply_token");
     deepEqual(req.body.messages, [testMsg]);
-    deepEqual(res, {});
+    equal(res["x-line-request-id"], "X-Line-Request-Id");
   });
 
   it("push", async () => {
@@ -65,7 +63,7 @@ describe("client", () => {
     equal(req.method, "POST");
     equal(req.body.to, "test_user_id");
     deepEqual(req.body.messages, [testMsg]);
-    deepEqual(res, {});
+    equal(res["x-line-request-id"], "X-Line-Request-Id");
   });
 
   it("multicast", async () => {
@@ -81,7 +79,17 @@ describe("client", () => {
       "test_user_id_3",
     ]);
     deepEqual(req.body.messages, [testMsg, testMsg]);
-    deepEqual(res, {});
+    equal(res["x-line-request-id"], "X-Line-Request-Id");
+  });
+
+  it("broadcast", async () => {
+    const res = await client.broadcast([testMsg, testMsg]);
+    const req = getRecentReq();
+    equal(req.headers.authorization, "Bearer test_channel_access_token");
+    equal(req.path, "/message/broadcast");
+    equal(req.method, "POST");
+    deepEqual(req.body.messages, [testMsg, testMsg]);
+    equal(res["x-line-request-id"], "X-Line-Request-Id");
   });
 
   it("getProfile", async () => {
@@ -361,5 +369,74 @@ describe("client", () => {
     equal(req.path, "/message/delivery/multicast");
     equal(req.query.date, date);
     equal(req.method, "GET");
+  });
+
+  it("getTargetLimitForAdditionalMessages", async () => {
+    await client.getTargetLimitForAdditionalMessages();
+    const req = getRecentReq();
+    equal(req.headers.authorization, "Bearer test_channel_access_token");
+    equal(req.path, "/message/quota");
+    equal(req.method, "GET");
+  });
+
+  it("getNumberOfMessagesSentThisMonth", async () => {
+    await client.getNumberOfMessagesSentThisMonth();
+    const req = getRecentReq();
+    equal(req.headers.authorization, "Bearer test_channel_access_token");
+    equal(req.path, "/message/quota/consumption");
+    equal(req.method, "GET");
+  });
+
+  it("getNumberOfSentBroadcastMessages", async () => {
+    const date = "20191231";
+    await client.getNumberOfSentBroadcastMessages(date);
+    const req = getRecentReq();
+    equal(req.headers.authorization, "Bearer test_channel_access_token");
+    equal(req.path, "/message/delivery/broadcast");
+    equal(req.query.date, date);
+    equal(req.method, "GET");
+  });
+
+  it("getNumberOfMessageDeliveries", async () => {
+    const date = "20191231";
+    await client.getNumberOfMessageDeliveries(date);
+    const req = getRecentReq();
+    equal(req.headers.authorization, "Bearer test_channel_access_token");
+    equal(req.path, "/insight/message/delivery");
+    equal(req.query.date, date);
+    equal(req.method, "GET");
+  });
+  it("getNumberOfFollowers", async () => {
+    const date = "20191231";
+    await client.getNumberOfFollowers(date);
+    const req = getRecentReq();
+    equal(req.headers.authorization, "Bearer test_channel_access_token");
+    equal(req.path, "/insight/followers");
+    equal(req.query.date, date);
+    equal(req.method, "GET");
+  });
+  it("getFriendDemographics", async () => {
+    await client.getFriendDemographics();
+    const req = getRecentReq();
+    equal(req.headers.authorization, "Bearer test_channel_access_token");
+    equal(req.path, "/insight/demographic");
+    equal(req.method, "GET");
+  });
+  it("fails on construct with no channelAccessToken", () => {
+    try {
+      new Client({ channelAccessToken: null });
+      ok(false);
+    } catch (err) {
+      equal(err.message, "no channel access token");
+    }
+  });
+
+  it("fails on pass non-Buffer to setRichMenu", async () => {
+    try {
+      await client.setRichMenuImage("test_rich_menu_id", null);
+      ok(false);
+    } catch (err) {
+      equal(err.message, "invalid data type for postBinary");
+    }
   });
 });